FIX : minseonglove 리뷰 반영
