+++ BidderBidder/LoginViewController.swift
@@ -81,19 +81,3 @@ class CornerButton: UIButton {
     
 }
 
-extension UIColor {
-    
-    static let skyBlueColor = UIColor(rgb: 0x21CEFF).cgColor
-    
-    convenience init(red: UInt8, green: UInt8, blue: UInt8) {
-        self.init(red: CGFloat(red) / 255.0, green: CGFloat(green) / 255.0, blue: CGFloat(blue) / 255.0, alpha: 1.0)
-    }
-    
-    convenience init(rgb: Int) {
-        self.init(
-            red: UInt8((rgb >> 16) & 0xFF),
-            green: UInt8((rgb >> 8) & 0xFF),
-            blue: UInt8(rgb & 0xFF)
-        )
-    }
-}
